requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
scipy==1.11.4
matplotlib==3.8.2
pyyaml==6.0.1
tqdm==4.66.2
# Optional but preferred (auto fallback provided if unavailable at runtime)
sentence-transformers==2.5.1
torch==2.2.2
umap-learn==0.5.5

joblib==1.3.2
